wrote tests for size_type constructor
