+++ tests/include/test_functions_header.hpp
@@ -12,6 +12,6 @@ void    assert(bool condition, const std::string& tagline = "");
 void    test_cleanup(void);
 void    test_emptiness(void);
 void    test_typedefs(void);
-void    test_vector_creation(void);
+void    test_vector_construction(void);
 
 #endif
