+++ tests/main.cpp
@@ -1,7 +1,7 @@
 #include "test_functions_header.hpp"
 
 int main() {
-    test_vector_creation();
+    test_vector_construction();
     test_cleanup();
     test_emptiness();
     test_typedefs();
